move category filter to components
